update test files
